Added a ContractResolver to the json deserializer to allow it to set non public members (such as Simple2DGrid.Bounds).
Added a new command line flag, -startpos x y to allow the robot's position to be set from the command line.

Moved the code for logging the robot's last position to the RobotInstance_TileEncountered event callback so the tile encountered could be more accurately determined.

Changed some of the logic in main to allow for loading a grid from file without   having to enter CommandLineMode.